feat(MATCH): adds close match sort functionality (#27)
* feat: Adds close match sort functionality

If the words fall below acronym for matching, it ranks how close the match string is to the test
string. It will return a number between 1 and 2 where 2 is a perfet match and 1 is no match. It does
this by dividing 1 by the difference of the spread of the characters in the match string within the
test string and the length of the match string and subtracting that quotient from 1. It then takes
that differneces and subtracts it from 2.

* Updates README, contributors, comments, and places
close match logic within the match ranking

* Fixes rank/number comments

* fixes contributors and comments

* Corrects rankings.MATCH comments to rankings.MATCHES

* fixes the close sort and the ignore diacritics test

* fixes long comment on line 159

* fixes comment on close match rank

* cleans up directory and resets dependencies

* updates the README for close match examples

* corrects the README close match example